backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dateutil==2.9.0.post0
python-multipart==0.0.17
itsdangerous==2.2.0
